Upload Protocol working
Client is able to connect and send messages to each machine
independently. The IP addresses are loaded from a configuration file,
which is placed in the same folder as those for bftsmart. However, the
file is "independent".

Proper controls have still to be implemented. This commit has only the
main structure were "some" data is exchanged.